refactor: replace 8 outer flow nodes with pipeline functions
Replace the outer node graphs (PocketFlow Node subclasses) with plain
async pipeline functions for both reflection and tool-use flows.

Reflection flow: 5 nodes (PrepareContextNode, TeXCountNode,
MediaExtractionNode, ResponseCycleNode, OutputNode) → single
executeReflectionRound() pipeline function. Round loop managed
directly in runReflectionFlow() instead of RoundPersistedFlow.

Tool-use flow: 3 nodes (ToolUsePrepareNode, ToolUseCycleNode,
ToolUseWaitNode) → three pipeline functions (prepareToolUse,
runToolUseCycle, waitForFollowUp). Cycle+wait loop managed
directly in runToolUseFlow() instead of PersistedFlow.

Inner cycle flows (ResponseCycleFlow, ToolUseCycleFlow) are
unchanged — they genuinely benefit from retry, branching, and
the node graph pattern.

Files deleted: 8 node classes (~1,060 LOC)
Files added: 2 pipeline modules (~820 LOC)
Net reduction: ~205 LOC + elimination of 6 bridge type definitions
