--- src/agent/implementations/flows/reflection/nodes/OutputNode.ts
@@ -1,275 +1,0 @@
-import { Node } from '@agent/node';
-import type { RoundFileMapping } from '@agent/output/types';
-import type { LatexDiffManager } from '@agent/output/LatexDiffManager';
-import { hasRoundOutputs, getStorageKey } from '@agent/output/outputState';
-import { extractFilesFromXml } from '@agent/output/xmlExtraction';
-import { traceFileLineage } from '@agent/output/lineageMapping';
-import { checkExpectedOutputs } from '@agent/output/outputValidation';
-import {
-  summarizeRound,
-  getRoundOutput,
-  type RoundSummary,
-} from '@agent/output/roundSummary';
-import { FlowTransition } from '@agent/core/flows/FlowTransitions';
-import { toErrorMessage } from '@common/errors';
-import { openBuildDisplayIfTex } from '@frontend/latex/openBuild';
-import { showInstructionWithSuppress } from '@frontend/ui/instruction';
-import type { AgentFileLocation, FileLocation } from '@utils/files';
-import { flexibleFS } from '@utils/files';
-import { bus } from '@eventBus/ProgressEventBus';
-import type { RoundOutput } from '@shared/schemas';
-
-import type { ReflectionFlowShared } from '../ReflectionFlowState';
-import type {
-  ReflectionFlowParams,
-  ReflectionServices,
-} from '../ReflectionServices';
-
-interface OutputPrepInput {
-  outputLocation: AgentFileLocation;
-  currentRound: number;
-  endTurn: boolean;
-}
-
-interface OutputExecResult {
-  roundOutput: RoundOutput;
-  summary: RoundSummary;
-}
-
-/** Execute an operation that can fail gracefully (logs warnings, doesn't throw). */
-async function tryOperation(
-  label: string,
-  operation: () => Promise<void>,
-  logger: { warn: (msg: string) => void },
-): Promise<void> {
-  try {
-    await operation();
-  } catch (error) {
-    logger.warn(`${label} failed: ${toErrorMessage(error)}`);
-  }
-}
-
-export class OutputNode<C = unknown> extends Node<
-  ReflectionFlowShared,
-  ReflectionFlowParams,
-  ReflectionServices<C>
-> {
-  async prep(shared: ReflectionFlowShared): Promise<OutputPrepInput> {
-    if (!shared.outputLocation) {
-      throw new Error(
-        'Output location not set - ResponseCycleNode must run first',
-      );
-    }
-
-    return {
-      outputLocation: shared.outputLocation,
-      currentRound: shared.currentRound,
-      endTurn: shared.endTurn,
-    };
-  }
-
-  async exec(prepRes: OutputPrepInput): Promise<OutputExecResult> {
-    const {
-      outputState,
-      xmlManager,
-      diffManager,
-      setting,
-      logger,
-      baseFiles,
-      shouldEnsureXmlStructure: shouldEnsureXml,
-    } = this.services;
-    const { outputLocation, currentRound, endTurn } = prepRes;
-
-    let mapping: RoundFileMapping | undefined;
-
-    // Only process if turn ended (model completed response)
-    if (endTurn) {
-      logger.debug(`Processing output for round ${currentRound}`);
-
-      if (shouldEnsureXml) {
-        await tryOperation(
-          'XML structure',
-          () =>
-            xmlManager.ensureCorrectXmlStructure(
-              outputLocation,
-              setting.documentTag,
-            ),
-          logger,
-        );
-      }
-
-      await tryOperation(
-        'Output processing',
-        () =>
-          extractFilesFromXml(
-            outputState,
-            this.services,
-            xmlManager,
-            outputLocation,
-            currentRound,
-          ),
-        logger,
-      );
-
-      if (hasRoundOutputs(outputState, currentRound)) {
-        mapping = traceFileLineage(outputState, baseFiles, currentRound);
-
-        await tryOperation(
-          'Latexdiff',
-          () =>
-            this.handleLatexdiff(
-              currentRound,
-              baseFiles,
-              mapping!,
-              diffManager,
-            ),
-          logger,
-        );
-      }
-    }
-
-    // Summarize round (pure data — no events)
-    const summary = await summarizeRound(
-      outputState,
-      this.services,
-      outputLocation,
-      currentRound,
-      { endTurn, mapping, isRewrite: setting.isRewrite },
-    );
-
-    // Get round output — critical, throw if it fails
-    const roundOutput = await getRoundOutput(
-      outputState,
-      baseFiles,
-      currentRound,
-      { isRewrite: setting.isRewrite },
-    );
-
-    return { roundOutput, summary };
-  }
-
-  async execFallback(
-    prepRes: OutputPrepInput,
-    error: Error,
-  ): Promise<OutputExecResult> {
-    const { logger, outputState, setting } = this.services;
-    const { outputLocation, currentRound, endTurn } = prepRes;
-    logger.warn(`Output processing failed: ${error.message}`);
-
-    // Still summarize what we can for post() side effects
-    let summary: RoundSummary;
-    try {
-      summary = await summarizeRound(
-        outputState,
-        this.services,
-        outputLocation,
-        currentRound,
-        { endTurn, isRewrite: setting.isRewrite },
-      );
-    } catch {
-      summary = {
-        storageKey: getStorageKey(outputState),
-        currRound: currentRound,
-        fileInfos: [],
-        filesToOpen: [],
-        outputFile: outputLocation,
-        endTurn,
-      };
-    }
-
-    return {
-      roundOutput: {
-        round: currentRound,
-        rawOutput: null,
-        outputs: [],
-        xmlSummary: {
-          tagContents: {},
-          documents: [],
-          singleOutputFile: null,
-          sourceLocation: null,
-        },
-      },
-      summary,
-    };
-  }
-
-  async post(
-    shared: ReflectionFlowShared,
-    prepRes: OutputPrepInput,
-    execRes: OutputExecResult,
-  ): Promise<string | undefined> {
-    const { streamId, logger, outputState } = this.services;
-    const { outputLocation, currentRound, endTurn } = prepRes;
-    const { summary, roundOutput } = execRes;
-
-    // Emit output files event
-    bus.emit('addOutputFiles', {
-      streamId,
-      storageKey: summary.storageKey,
-      filesByRound: { [currentRound]: summary.fileInfos },
-    });
-
-    // Open files that haven't been opened yet
-    for (const location of summary.filesToOpen) {
-      await tryOperation(
-        `Open file ${location.absolutePath}`,
-        () => openBuildDisplayIfTex(location, { preserveFocus: true }),
-        logger,
-      );
-    }
-
-    // Validate expected outputs if turn ended
-    if (endTurn) {
-      await tryOperation(
-        'Validate expected outputs',
-        async () => {
-          const validationResult = await checkExpectedOutputs(
-            outputState,
-            this.services,
-            outputLocation,
-            currentRound,
-            summary.stage,
-          );
-
-          bus.emit('updateMissingOutputs', {
-            streamId,
-            storageKey: validationResult.storageKey,
-            filesByRound: { [currentRound]: validationResult.missing },
-          });
-
-          if (validationResult.missing.length > 0) {
-            await showInstructionWithSuppress(
-              'missingOutputsInfo',
-              'Missing output files detected',
-            );
-          }
-        },
-        logger,
-      );
-    }
-
-    // Store round output
-    shared.roundOutputs[currentRound] = roundOutput;
-
-    return FlowTransition.DEFAULT;
-  }
-
-  private async handleLatexdiff(
-    currentRound: number,
-    baseFiles: FileLocation[],
-    mapping: RoundFileMapping,
-    diffManager: LatexDiffManager,
-  ): Promise<void> {
-    const { logger } = this.services;
-
-    const existingBase = await Promise.all(
-      baseFiles.map((f) => flexibleFS.exists(f)),
-    );
-    if (!existingBase.some(Boolean)) {
-      logger.debug('No base files found for latexdiff');
-      return;
-    }
-
-    await diffManager.handleLatexdiffofOutput(currentRound, mapping);
-  }
-}
